+++ test/specs/languages/test-typescript.el
@@ -38,7 +38,23 @@
   \"failure\":\"missing semicolon\",
   \"name\":\"sample.ts\",
   \"ruleName\":\"semicolon\",
-  \"startPosition\":{\"character\":14,\"line\":2,\"position\":76}}]"))
+  \"startPosition\":{\"character\":14,\"line\":2,\"position\":76}}]")
+          (json-with-deprecations "no-unused-variable is deprecated. Use the tsc compiler options --noUnusedParameters and --noUnusedLocals instead.
+
+  Could not find implementations for the following rules specified in the configuration:
+      label-undefined
+      no-constructor-vars
+      no-duplicate-key
+      no-unreachable
+      use-strict
+  Try upgrading TSLint and/or ensuring that you have all necessary custom rules installed.
+  If TSLint was recently upgraded, you may have old rules configured which need to be cleaned up.
+
+[{\"endPosition\":{\"character\":25,\"line\":0,\"position\":25},
+ \"failure\":\"unused variable: 'invalidAlignment'\",
+ \"name\":\"sample.ts\",
+ \"ruleName\":\"no-unused-variable\",
+ \"startPosition\":{\"character\":9,\"line\":0,\"position\":9}}]"))
       (it "parses TSLint JSON output"
         (expect (flycheck-parse-tslint json 'checker 'buffer)
                 :to-be-equal-flycheck-errors
@@ -52,6 +68,16 @@
                  (flycheck-error-new-at 3 15 'warning
                                         "missing semicolon"
                                         :id "semicolon"
+                                        :checker 'checker
+                                        :buffer 'buffer
+                                        :filename "sample.ts"))))
+      (it "parses TSLint JSON output with deprecation output"
+        (expect (flycheck-parse-tslint json-with-deprecations 'checker 'buffer)
+                :to-be-equal-flycheck-errors
+                (list
+                 (flycheck-error-new-at 1 10 'warning
+                                        "unused variable: 'invalidAlignment'"
+                                        :id "no-unused-variable"
                                         :checker 'checker
                                         :buffer 'buffer
                                         :filename "sample.ts")))))))